refactor(mcp/middleware): flatten mcp/middleware/__tests__ — rename 4 parallel suites (#2565 slice 8)
mcp/middleware/__tests__ had 4 files with basename collisions:

- middleware-chain.test.ts (top:1387 / nested:377)
- request-context.test.ts (top:314 / nested:183)
- secure-handler.test.ts (top:1425 / nested:256)
- tool-rate-limiter.test.ts (top:219 / nested:255)

Each nested file has a single wrapping describe block (Middleware
Chain, RequestContext, SecureHandler, ToolRateLimiterFactory) at the
top level — pattern matches the parallel-suite style from earlier
slices. Top-level files are larger and more comprehensive in 3 of 4.

Use the consistent -extended.test.ts suffix to eliminate basename
collisions while preserving all tests. Fix imports. Remove the empty
__tests__/ directory.

All 616 mcp/middleware tests still pass.
